Create tests for pseudo
